recommender-fastapi/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sentence-transformers==2.2.2
numpy==1.24.3
python-multipart==0.0.6
pymongo==4.6.0
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
huggingface-hub==0.19.4
torch==2.1.0
transformers==4.35.2
datasets==2.14.6
accelerate==0.24.1
